Fixed bug in DataOrderTable, moved defines into header.
There was a bug in the declaration of DataOrderTable, where it was defined as 24x3 when it's actually 24x4. I am really surprised I haven't noticed this/that the code worked at all.  I guess all the pokemon I have tested on has orders that were in the low end.