Add game selector to UI
Add game selector so the user can select
from which Gen3 game he is loading a RAM
file for.  This will allow us to easily
find where in the file the pokemon's data
is contained.

Note that this does not do anything yet!
It is simply a selector on the UI not
hooked up to anything.