Add save parser to parse .sav files
Instead of parsing gzipped VBA-only
ram snapshots we will parse .sav files.
This requires data the data in 3968B
chunks and re-ordering them.

Save format is as follows:
32 blocks
14 blocks for save slot 1
14 blocks for save slot 2
4 blocks for Hall of Fame
each block has 3968B of data, followed
by 116B of padding, then a 12B footer.
The footer has:
1 byte block number (0..13)
1 byte padding
2 byte checksum
4 byte validation = 0x08012025
4 byte savenumber

There will only be two savenumbers per
save file.  The higher of the two is
what the game loads.

Checksum is generated by adding up all
the data, 4 bytes as a time, adding the
upper 16 bits of the result to the bottom
16 and then casting to two bytes.

THIS CHANGE WILL BREAK THE GUI-SIDE!