GUI: Populate combo boxes using same list.
Popular Item, Species, and Attack combo boxes using the
list of names instead of having a copy of the list. The
previous list of items was also incomplete while the
list of species was incorrect.

Also only have two choices for the game, as FR/LG have
the same offset for the pokemon belt just as R/S/E do.